fix(api): serve /api/init from a committed seeded DB; fast paginated fallback
- Ship the seeded 1981-2026 season database (f1_data.db, previously
  gitignored): every /api/init?year= now serves from SQLite in <0.5s.
  Before this, Render deployed with no data, so every year fell into a
  per-round Jolpica fallback (~24 sequential requests + sleeps, ~96s
  observed) that occupied both gunicorn workers and starved every other
  endpoint - the Season Simulator's "very long load".
- Rewrite the fallback to bulk-paginate Jolpica results, merging rows by
  round across the 100-row page cap (verified 6.2s for a full unseeded
  season, then served from the in-memory cache).
- Normalize the SQLite /api/init payload to the exact shape the frontend
  consumes: teams/drivers keyed id/team, standings driverId/teamId,
  raceResults keyed "<year>_r<round>" like the schedule. The old DB-path
  keys (constructorId, entityId, bare round numbers) never shipped
  because the DB never shipped; committing it would have activated them.
- seed_data.py: parse standings points as float - half-points seasons
  (1984, 1991, 2009: int('71.5')) crashed the seeder. standings.points
  is now NUMERIC(6,1) in both SQLite and PG schemas; 46/46 seasons seed
  cleanly.

+++ flask-app/seed_data.py
@@ -302,7 +302,9 @@ def _seed_standings(year: int, all_results: dict, constructor_map: dict):
             for s in standings[0].get("DriverStandings", []):
                 did = normalize_driver_id(s.get("Driver", {}).get("driverId", ""))
                 pos = int(s.get("position", 0))
-                pts = int(s.get("points", 0))
+                # Half-points seasons exist (e.g. 1984 Prost 71.5 after the
+                # half-points Monaco GP) — parse as float, never int().
+                pts = float(s.get("points", 0))
                 if did and pos > 0:
                     insert_standing(year, did, "driver", pos, pts)
 
@@ -314,7 +316,7 @@ def _seed_standings(year: int, all_results: dict, constructor_map: dict):
             for s in standings[0].get("ConstructorStandings", []):
                 tid = s.get("Constructor", {}).get("constructorId", "")
                 pos = int(s.get("position", 0))
-                pts = int(s.get("points", 0))
+                pts = float(s.get("points", 0))
                 if tid and pos > 0:
                     insert_standing(year, tid, "constructor", pos, pts)
